Добавление сервисов для госпитализации

See merge request maksimovoleg19/FMH!4

+++ src/main/java/ru/iteco/fmh/dao/repository/ClaimRepository.java
@@ -3,5 +3,5 @@ package ru.iteco.fmh.dao.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.iteco.fmh.model.Claim;
 
-public interface ClaimRepository extends JpaRepository<Claim, Long> {
+public interface ClaimRepository extends JpaRepository<Claim, Integer> {
 }
